+++ admin/api/products/edit_product.php
@@ -1,5 +1,5 @@
 <?php
-require_once "pdo-connect.php";
+require_once __DIR__ . "../../../../camping_db.php";
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $id = $_POST['id'];
@@ -13,7 +13,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     header('Content-Type: application/json; charset=utf-8');
 
     try {
-        $stmt = $pdo->prepare("UPDATE products SET name = ?, price = ?, stock = ?, status = ? ,category_id = ?,subcategory_id = ? WHERE id = ?");
+        $stmt = $db->prepare("UPDATE products SET name = ?, price = ?, stock = ?, status = ? ,category_id = ?,subcategory_id = ? WHERE id = ?");
         $stmt->execute([$name, $price, $stock, $status, $category_id, $subcategory_id, $id]);
 
         echo json_encode(["success" => true, "message" => "更新成功"]);
